+++ index.html
@@ -40,6 +40,6 @@
         </ul>
       </div>
     </main>
-    <script src="./script.js"></script>
+    <script type="module" src="./script.js"></script>
   </body>
 </html>
