+++ script.js
@@ -1,6 +1,12 @@
 /*immediately invoked function expression o bien por sus siglas: IIFE, 
 que es tal cual, son funciones que en cuanto se declaran, se ejecutan*/
-(()=>{
+/* (()=>{
+
+    esto tambien se previene utilizando el esquema de modulos
+    imoprtante: se debe cambiar agregar a la etiqueta script del HTML el tipe='module'
+ */
+import checkComplete from "./components/checkComplete.js";
+import deleteIcon from "./components/deleteIcon.js";
 
 const btn = document.querySelector('[data-form-btn]')
 
@@ -13,38 +19,17 @@ const createTask = (evento) => {
     task.classList.add('card')
     input.value = '';
     const taskContent = document.createElement('div')
-    taskContent.appendChild(checkComplete())
     const titleTask = document.createElement('span')
     titleTask.classList.add('task')
     titleTask.innerText = value;
+    taskContent.appendChild(checkComplete())
     taskContent.appendChild(titleTask)
-    const content = `
-        <i class="fas fa-trash-alt trashIcon icon"></i>`
     //task.innerHTML = content;
     task.appendChild(taskContent)
+    task.appendChild(deleteIcon())
     list.appendChild(task)
 }
 
 btn.addEventListener('click', createTask)
 
-const checkComplete = ()=>{
-    const i = document.createElement('i');
-    i.classList.add('far','fa-check-square','icon');
-    i.addEventListener('click',completeTask)
-    return i;
-}
-
-const completeTask = (evento)=>{
-    const element = evento.target
-    /*  
-    element.classList.remove('far')
-    element.classList.add('fas');
-    element.classList.add('completeIcon') 
-    con TOGGLE podemos verificar: si una clase no esta => se agrega... de lo contrario se quita
-    */
-    element.classList.toggle('fas');
-    element.classList.toggle('far');
-    element.classList.toggle('completeIcon') 
-}
-
-})()
+/* })() */
